+++ NEWS.md
@@ -2,3 +2,4 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 * Fixes due to changes in upstream trending package.
+* `evaluate_resampling()` now defaults to 10-fold cross-validation instead of Leave-One-Out cross-validation.